Count down timer and circle progress bar
The countdown timer was changed so that it would count down the time in
minutes, it is able to count down up to about 24 days using a dd:hh:mm
format where if there is less then a day it day section disappears.

The circle progress bar was just reworked so that I worked with the same
time frame that the timer was working on.

Both are currently using item 0 that it pull from parse.